+++ changelog.md
@@ -1,6 +1,8 @@
 ## Version 0.8.0
  - Custmizable homepage / theme colors
+ - Create smaller subnets when using force secure / links to not hit IP range limit
  - Stop showing Docker not connected when first loading status
+ - Self-heal containers that have lost their network configurations
  - Add privacy settings to external links
 
 ## Version 0.7.1 -> 0.7.10